# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CalOS Multi-LLM Router API
  version: 1.1.0
  description: |
    CalOS Agent Router - AI-powered routing system with bot detection, multi-provider LLM support, and reputation-based rate limiting.

    ## Features
    - 🤖 **4 LLM Providers**: OpenAI (GPT-4), Anthropic (Claude), DeepSeek, Ollama (local)
    - 🛡️ **Bot Detection**: Zero-knowledge proof-of-personhood without KYC
    - 🔐 **Soulfra Identity**: Self-sovereign cryptographic identity
    - 📊 **Smart Routing**: Automatic provider selection based on task type
    - ⚡ **Rate Limiting**: Reputation-based tiered limits
    - 💰 **Cost Optimization**: Prefer free/cheap providers when possible

    ## Authentication Flow
    1. `POST /api/llm/request-access` - Get challenge
    2. Create Soulfra identity, solve proof-of-work
    3. `POST /api/llm/verify-personhood` - Get access token
    4. Use access token in `Authorization: Bearer <token>` header

    ## Domain Contexts (Task Types)
    - `code`: Programming, debugging, code review
    - `creative`: Writing, stories, creative content
    - `reasoning`: Complex analysis, logic, explanations
    - `fact`: Simple factual queries
    - `simple`: Basic questions

  contact:
    name: CalOS Development
    url: https://github.com/calos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.calos.ai
...
paths:
...
  /api/llm/complete:
    post:
      tags:
        - LLM Operations
      summary: Complete a prompt
      description: |
        Send a prompt to the multi-LLM router. The router automatically selects the best provider based on:
        - Task type (code, creative, reasoning, fact, simple)
        - Routing strategy (smart, cheapest, fastest, best-quality)
        - Provider availability
        - Cost optimization

        **Task Type Routing:**
        - `code`: Ollama CodeLlama > GPT-4 > Claude
        - `creative`: Claude > GPT-4 > DeepSeek
        - `reasoning`: DeepSeek Reasoner > GPT-4 > Claude
        - `fact`: Ollama > DeepSeek > GPT-3.5
        - `simple`: Ollama (free and fast)
      operationId: complete
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletionRequest'
      responses:
        '200':
          description: Completion successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied (blacklisted or session expired)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: LLM error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token obtained from `/api/llm/verify-personhood` endpoint.
        Include in Authorization header: `Authorization: Bearer <access_token>`
  schemas:
...
    CompletionRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          description: The prompt to complete
          example: "Explain quantum computing in simple terms"
        systemPrompt:
          type: string
          description: System prompt (optional)
          example: "You are a helpful assistant."
        taskType:
          type: string
          enum: [code, creative, reasoning, fact, simple]
          description: Task type for smart routing
          example: "reasoning"
        maxTokens:
          type: integer
          description: Maximum tokens to generate
          default: 1000
          example: 500
        temperature:
          type: number
          format: float
          description: Sampling temperature (0.0 to 2.0)
          default: 0.7
          example: 0.7
        preferredProvider:
          type: string
          enum: [openai, anthropic, deepseek, ollama]
          description: Force specific provider (optional)
        model:
          type: string
          description: Force specific model (optional)
          example: "gpt-4"
    CompletionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        response:
          type: object
          properties:
            text:
              type: string
              example: "Quantum computing is..."
            provider:
              type: string
              enum: [openai, anthropic, deepseek, ollama]
              example: "ollama"
            model:
              type: string
              example: "llama3.2:3b"
            usage:
              type: object
              properties:
                prompt_tokens:
                  type: integer
                  example: 15
                completion_tokens:
                  type: integer
                  example: 200
                total_tokens:
                  type: integer
                  example: 215
            finishReason:
              type: string
              enum: [stop, length]
              example: "stop"
            latency:
              type: integer
              description: Response time in milliseconds
              example: 2893
        session:
          type: object
          properties:
            identityID:
              type: string
            tier:
              type: string
            requestCount:
              type: integer
        rateLimit:
          type: object
          properties:
            remaining:
              type: object
              properties:
                hourly:
                  type: integer
                daily:
                  type: integer
            limits:
              type: object
              properties:
                hourly:
                  type: integer
                daily:
                  type: integer
            resetAt:
              type: object
              properties:
                hourly:
                  type: string
                  format: date-time
                daily:
                  type: string
                  format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: "invalid_request"
        message:
          type: string
          description: Human-readable error message
          example: "Missing required field: prompt"
        reason:
          type: string
          description: Detailed error reason (optional)
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: "rate_limit_exceeded"
        reason:
          type: string
          enum: [hourly_limit_exceeded, daily_limit_exceeded]
          example: "hourly_limit_exceeded"
        limits:
          type: object
          properties:
            hourly:
              type: integer
            daily:
              type: integer
        remaining:
          type: object
          properties:
            hourly:
              type: integer
            daily:
              type: integer
        resetAt:
          type: object
          properties:
            hourly:
              type: string
              format: date-time
            daily:
              type: string
              format: date-time
        message:
          type: string
          example: "Rate limit exceeded. You can make 100 requests per hour."